Split loader from 2D action classification udf
